Remove charge cache in pattern parameters
Currently, we're storing the charge value of the pattern track
parameters in a cached value, but this is not really necessary.
Computing the charge is so trivial and fast (from the q/p value) that
there really is no reason to cache this. We use the lazy binding
behaviour of the charge method to simply recompute on the fly.