PanTau code cleanup
Hello,

This MR is cleaning up the PanTau code, which is no longer maintained but still used in the tau reconstruction.
This is just to improve code readability for a possible rework or simplification of the code, which might be attemped before the reprocessing.
I also dropped a few calculations that are made internally but used nowhere.
This MR does not change the reconstruction output.

Cheers,
Bertrand