EventTagAlgs: delete obsolete TAG making package
TAG making ("doWriteTAG") is no longer supported and this package only
contains unused algorithms. Remove references from other packages to
this package or delete them entirely if obsolete.

The `EventTagGlobal.AttributeList` variable is still required in
RecExCommon. But in many other places we anyhow just use the hard-coded
value `"SimpleTag"`. So do the same here to avoid this trivial
dependency.