Removed callback registration from AmdcsimrecAthenaSvc
Given Muon alignment handling has been transitioned to condition algorithms,
the callback registration for fetching Muon alignments became redundant.
With this change AmdcsimrecAthenaSvc gets all required information from
the MuonDetectorManager at initialization.

If in the future somebody desires to make AmdcsimrecAthenaSvc sensitive
to Muon alignment values, then a substantial redesign of this service,
and other services that depend on it, will be required.

As a bonus, with this change we drop the last remaining client of the I
GeoModelSvc::align() function, such that the latter can now be removed.