Switch muon trigger to not use ROB based RPC PRD decoding
Since it looks like there is some non-obvious bug in the ROB based RPC PRD decoding (see ATR-22409), I'm temporarily switching the seeded decoding (used only by the trigger) to decode based on the list of hashIDs in an RoI instead of the list of ROBs until the bug can be better investigated (since this is hopefully a temporary change I'm leaving the ROB based decdoing lines commentd out in the decoding algorithm until).
This changes somewhat the muons we reconstruct (mostly in the early stages of the trigger), so the count reference files have been updated accordingly.