CaloTriggerTool: Trying to fix MT dictionary error.
CaloTriggerTowerService used to read LArTTCellMap via a conditions callback.
This is not thread-friendly, so it was changed; it is now retrieved from
the detector store during event processing.  However, this can lead
to the I/O system triggering the creation of TClass instances for the
LArTTCellMap persistent classes, and TClass has been observed to sometimes
malfunction when created in a MT context (at least as of root 6.22.00).
The resulting dictionary failures were causing jobs to occasionally
crash with bad_alloc exceptions.  Try to fix by forcing the needed
TClass instances to be created during initialize.