MuonSTGC_CnvTools: const fixes
Preparing to make methods of IMuonRdoToPrepDataTool const.

Adjust sTgcRdoToPrepDataTool* so that its interfaces may be made const.
Also enable thread-safety checking.
The old pre-MT sTgcRdoToPrepDataTool is declared as not thread-safe,
as it has some mutable state, while the mutable state is removed from
sTgcRdoToPrepDataToolMT.  sTgcRdoToPrepDataToolCore is now abstract
and is removed from the entries list.