Avoid an early creation of CaloDescriptors.
The CaloDescriptors must not be created before the alignment callback has been
fired, because the descriptors cache the alignment transform at construction
time.

The creation of CaloDescriptors is e.g. triggered by the retrieval of the
calorimeter detector description manager. To ensure that this happens after the
alignment callback they are not retrieved anymore during the initialization of
the calorimeter volume builders.