Mostly changes for "by leg" reporting of jets passing het hypo - see ATR-20700
This MR also contains

- removal of the shared leaf calculation from the node tree,
and FastReducer C++ code. For now we will not have sharing
in a tree, but will be building forests made of independent trees.
This will allow maximum sharing across trees.
     ConditionsToolSetterFastReduction.py
     FastReducer.cxx, h
     TrigJetHypoToolConfig_fastreduction.cxx, h
     groupsMatcherFactoryMT.cxx, h

- add leg labels (eg "leg000") to all leaf nodes
      chainDict2jetLabel.py

- collect passing jets by leg label
  xAODJetCollector.h, cxx

- Easier debugging output
  ICapacityCheckedCondition.h

- update reference file: for passing events, vbenf was
reporting all jets passing leaf cuts. This list is reduced
to those that also pass the root node.
   Trigger/TrigValidation/TriggerTest/share/ref_data_v1Dev_build.re