TrigAnalysisTest: Remove obsolete script
The script `test_trigAna_RDOtoRDOTrig_mt1_build.py` was renamed to `test_trigAna_RDOtoRDOTrig_v1Dev_build.py` but a hacky link with the old name to the new one was needed temporarily because the old name was present in CI configuration. Since the CI configuration was updated, the "link" script is obsolete and can be removed.