switch AsgTools from IJobOptionsSvc to IOptionsSvc
This should address ATLASG-1551, though I didn't perform a complete
migration.  Doing private sub-sub-tools with the AnaToolHandle in
Athena doesn't seem to work correctly anymore.  Given that
AnaToolHandle is no longer the recommended way of creating sub-tools
anymore I'd rather just see if this breaks anything.  If it does I'll
decide whether to try and fix the AnaToolHandle implementation, or
migrate the users to AsgToolConfig instead.