MadGraphControl Updating printing to make things a bit more clear
This re-orders some operations so that the call to print cards comes
only after all the setting manipulation has been done. This ensures that
the cards that are printed are really the ones with which we run.

Also removing printout from the consistency check function, which had
resulted in the settings being printed a second time unnecessarily.

Also cleaning up a setting in a SUSY common file, which was anyway being
overridden by the infrastructure (but now we won't have the warning).