Move the TruthCategoriesDecorator to release 22
This tool runs during derivation and adds Higgs STXS truth information to the AOD.
To run it requires the TruthConverters package that was not yet in the release.
Additionally I had to make a couple of modifications to the HiggsTruthCategoryTool which is only used by this tool.
I was able to successfully run on a Higgs AOD with this enabled as part of PHYS and the results looked reasonable.