Each stream writes to seperate file
Setting multiple streams to write to the same file causes problems when
there is an overlap in the streams' metadata itemlists that is not
1-to-1: those items in the overlap will add more entries to the metadata
tree than the others causing a failure when attempting to close the file
after writing. If the list is 1-to-1 overlapping the metadata tree has
multiple entries, which may lead to undefined behaviour.

There is no reason to expect multiple output streams to need to write to
the same file. We do not support that configuration, and should not test
for it.