METReconstruction: calculate dR directly
The MET reconstruction need to retrieve the clusters around 0.2 cone of
the tau candidate. Since the vertex of the tau candidate could be
different from PV0, a vertex correction is needed. The corrected
clusters are decorated onto the tau candidate, but is transient only.
This break the MET reconstruction from ESD/AOD. After recent
development of tau reconstruction, the correction is now much simple,
and we could do it directly in METReconstruction.

This should fix ATLASRECTS-5793.