tauRec: drop BDT tau ID (ATLTAU-1719)
Hello,

This MR is dropping the calculation of BDT tau ID, which was superseded by the RNN ID (in R21 already).
The trigger monitoring should no longer rely on offline BDT tau ID (!38174 , !38734 ).
Same for the PHYSVAL monitoring (!38911 ).
To be checked for the DQ monitoring.

The change in reconstruction output looks as expected, e.g.:
```
002.TauJetsAux.isTauFlags.1 4030464000 -> 4026531840 => diff= [0.02440215%]
002.TauJetsAux.BDTJetScore.0 -0.8960119485855103 -> 0.0 => diff= [50.00000000%]
002.TauJetsAux.BDTJetScore.1 0.9878825545310974 -> 0.0 => diff= [50.00000000%]
002.TauJetsAux.BDTJetScoreSigTrans.0 0.002360016806051135 -> 0.0 => diff= [50.00000000%]
002.TauJetsAux.BDTJetScoreSigTrans.1 0.8536615967750549 -> 0.0 => diff= [50.00000000%]
```

Cheers,
Bertrand