Refactor code to read GeoDB
Move detector-specific code to read the geometry database into
standalone "[Det]GeoDB.py" modules. Use this code both in the old
jobproperties as well as in the new-style `AutoConfigFlags` to avoid the
copy&paste of the same code.

Also fix the caching of `AutoConfigFlags.GetDetDescrInfo` to make sure the
geometry DB is only queried once per geometry tag.