Add unit tests for the ByteStreamMetadataTool
Test the functionality of the ByteStreamMetadataTool. The tool should do
nothing wihtout complaint for empty input. The tool should transcribes a
ByteStreamMetadata object from input into a container in output. The
tool should transcribe a ByteStreamMetadataContainer from input to
output. Should there already be content in the output, matching content
should be ignored and new content should be appended.