xAODTau: drop range check on eta and phi
Hello,

This MR is dropping a range check on tau eta and phi that was introduced because we used to set dummy default values (-1111) for eta and phi for 0-prong and >5-prong taus, leading to eta = NAN upon retrieval.
This has been fixed in the reconstruction (ATLTAU-1665), where we no longer set dummy values.
So we can now drop the range check in the TauJet interface class, to save a tiny amount of CPU.
This MR does not change the reconstruction output.
It can wait until 22.0.23 is built.

Cheers,
Bertrand