tauRec: add flags for calibration files
Hello,

This MR is adding tauRecFlags for the calibration files of the main tau algorithms, so that they can be changed on the fly with a preExec.
It will be very useful for the upcoming R22 tuning campaign, in case we need to rerun reconstruction with updated tunes.
A couple of deprecated flags were dropped. This MR does not change the reconstruction output.

Cheers,
Bertrand