Fix filling of monitoring histograms with kVec/kVecUO options.
Histograms with those options were not filled at all, there was a logic
bug in `VecHistogramFiller1D::fill` method. The fix is limited to that
class only and I have added a unit test for those two options. I also
noticed that HistogramDef contents is not initialized and I switched to
using MockHistogramDef in couple of other unit tests where I think it
may be important to have properly initialized definition. And I think
HistogramDef needs a default constructor which makes instances in a
known state (or per-member initializers).