Add the correct views for the Tau track collections
The Tau_IDTrig tracks cannot be retrieved from TauIso Roi, and the TauIso_FTF, nor the Tau_IDTrig
tracks can be retrieved frmo the TauIsoBDT Roi.

In the EDMConfig the relevant views are not in the lists for these track collections so these have been added.

There is a caveat that I am nt sure of the view name for the precisopn tracking
for the TauIsoBDT - there seems to be no dedicated PRecision tracking configuration
python function, so the  Tau_IDTrig collection may still end up as unretrievable.