CxxUtils: Add SimpleUpdater.
Add a simple implementation of the Updater interface.
It doesn't keep track of slots, and does not attempt to delete any garbage
until the SimpleUpdater object itself is deleted.

This can be appropriate for something like a ConcurrentStrMap if one
has a good guess in advance of the size of the table.