Speed up the primary vertex analysis for the ID Trigger validation
We run two vertex analyses - one in the standard offline processing,
and another with a primary vertex selection. This second processing
would previously also re-run all the other ttbar based analyses.

In the ttbar jobs, these take some considreable time, and are not needed,
as they have already been run in the standard job.

This adds a preprocessor macros to skip the confguration of the non-vertex
analysis chain for the primary vertex analysis