tauRecTools: protect against Log10(0) in very rare cases
Hello,

This MR adds a protection against computing Log10(0) for very rare cases (10 taus in 15 million events, as observed in gamma* MC).
It seems we can have "pathological" taus where clusters have FIRST_ENG_DENS = 0.
This seems to happen when |eta|~1.5, when all the cluster energy is in the presampler.
This MR does not change the reconstruction output except in very rare case, where xAOD::TauJetParameters::ClustersMeanFirstEngDens used to be -inf, and is now 0.

Cheers,
Bertrand