xAODTau: protection against tau track thinning
Hello,

This MR is mostly importing !26767 from 21.2 to master, with the original MR description pasted below.

A protection is added against invalid TauTrack links when retrieving the tau tracks or the number of tau tracks.
This is required to support tau track thinning, used e.g. in DAOD_PHYS.
Now, only the tau tracks that have a valid link are considered in the TauJets interface class.
In particular, the behaviour of nAllTracks() had to be changed to be consistent with allTracks().size().
If one only keeps e.g. classifiedCharged tau tracks by thinning away the other tracks, nAllTracks() will now return the number of classifiedCharged tracks.

Something new compared to the R21 MR: to keep track of the total number of tracks associated with the tau in the reconstruction, a new Aux variable is introduced (mapped to xAOD::TauJetParameters::nAllTracks). So the user is now able to tell if tracks were thinned in the AOD->DAOD step, by comparing tau->nAllTracks() to tau->detail(xAOD::TauJetParameters::nAllTracks, nTracksTot).

Cheers,
Bertrand