rework subtools in AsgComponentConfig (was wrap AnaToolHandle around AsgToolConfig)
Since I'm back-porting to 21.2 I didn't modify the AnaToolHandle
implementation, but I did update AsgComponentConfig accordingly.

Original commit message:

The basic reasoning here is that AnaToolHandle is essentially the
out-dated and discouraged mechanism for creating tools, and has been
replaced with AsgToolConfig for quite a while (or alternatively
indirectly via AnaAlgorithmConfig).

I did have to disable some Athena tests for setting sub-tools in
Athena.  There are already some of those disabled, and the basic logic
is that since this is the legacy way of doing things, unless this
breaks existing code (as verified by CI tests) I no longer want to
support all the possible edge cases in this rather brittle and complex
piece of code.