tauRecTools: drop unused tau track decorations
Hello,

This MR is dropping a few tau track decorations that are not needed (either redundant or not useful):

* TauJetsAuxDyn.nRnnChargedTracks is identical to TauJetsAux.nChargedTracks
* TauJetsAuxDyn.nRnnIsolationTracks is identical to TauJetsAux.nIsolatedTracks
* TauJetsAuxDyn.nRnnUnclassifiedTracks can be obtained from unitarity, from other tau track multiplicities (and in practice, is always equal to 0)
* TauTracksAuxDyn.rnn_unclassifiedScore can be obtained from unitarity, from other tau track RNN scores (and in practice, is always equal to 0)

nRnnConvertionTracks and nRnnFakeTracks decorations are kept for now (although renamed), as they don't exist as Aux variables, but this may be harmonised at some point.
The code has been streamlined a bit. A few unused input variables were commented out, we'll come back to this when updating the final list of inputs to be used for R22.

Except for the renamed and discontinued decorations, this MR does not change the reconstruction output.
With one exception: before, when the tau had no track, nRnnConvertionTracks and nRnnFakeTracks were not decorated onto the tau because of the early "return SUCCESS" in the classifyTracks method (this must have been overlooked).
Now, nConversionTracks and nFakeTracks are always decorated.

Cheers,
Bertrand