CaloClusterCorrection: Work around spurious zeros in rfac-v5.
The rfac-v5 interpolation tables contain a couple spurious zeros.
This will cause the rfac correction to blow up in the immediate
vicinity of these points; a FPE has also been observed when a cluster
exactly hits the zero point.

We don't want to put a new version in the databases at this point.
Instead, work around by replacing zeros in the rfac interpolation
table with an average of adjacent points.