tauRecTools: handle vertex correction in TauCalibrateLC
Hello,

This MR is fixing a small inconsistency in the calo-only tau energy scale calibration.
In the offline tau reconstruction, by default, the topocluster momentum is corrected for the tau vertex position.
However, this was not propagated to the calo-only energy calibration which currently uses a non-vertex-corrected pt as input (ptDetectorAxis).
This was overlooked in the R20->R21 migration when we switched from an energy correction to a pt correction.
This is now handled properly. The impact of the vertex correction on the tau pt is expected to be very small.
Yet, at the moment, it has no impact at all.
I suspect it comes from the fact that ptDetectorAxis and ptIntermediate are currently identical, which should not be.
To be investigated.

Cheers,
Bertrand