Convert mc_channel_number to string for TagInfo
In the /TagInfo object the MC channel number is stored as a string not
an integer. We must have missed this in the past with either the object
or the field not present. Now add converted string, should that throw
try adding as integer, should that still fail print error message.