Write output FileMetaData as soon as possible
Writing the output FileMetaData on preFinalize causes problems in
EventService: event ranges will be written out before the FileMetaData
object is created. This causes an error when the stream is told to write
an object that does not (yet) exist.

Enhancement: make the log messages from the FileMetaDataCreatorTool more
informative.