Avoid using BeginEvent Incident to Trigger TrackRecordCollection creation in ISF
This commit removes the dependency of the `EntryLayerTool` on IIncidentListener.
The old `handle(...)` method is replaced by `setupEvent`, which is called from
`ParticleBrokerDynamicOnReadIn::initializeEvent()`.
This change prevents the `TrackRecordCollection` objects from being created in the
eventStore in the case that ISF is not run. This is needed for the re-simulation
workflow, where some events are simply copied and some are re-simulated.