fix tau PHYSVAL monitoring crash
Hello,

This MR is adding a patch to the tau PHYSVAL monitoring.
It seems the `hasDiscriminant` method from xAODTau is not doing what it is supposed to do, namely, check whether a discriminant is present in the input file.
Instead, it checks whether the discriminant is present in TauJetAuxContainer_v3, which returns true for both R21 and R22 AODs, since they both use the v3 Aux version, while the RNN eVeto tool is only run in R22. We will have to fix the xAODTau EDM, but for now, I simply add a patch in the monitoring.

Cheers,
Bertrand