Fix for ATLASRECTS-5888
This ensures that the final segment collections written out contain both MuGirl and Muon(Moo)SegmentFinderAlg segments.
The xAOD segment container is now created and filled in the MuonCreatorAlg (so the xAODMuonCnvAlg is removed from the configuration).
The Muon(Moo)SegmentFinderAlg segments are copied into the final Trk::SegmentCollection in MuonCreatorAlg.
Conversion of the segments now happens in the the MuonCreatorAlg as well.
MuGirl segments are only added to the collection if they correspond to a muon with primary author MuGirl.
The MuonSegmentContainer used for associating segments to muons is now passed directly to the tool, rather than retrieved from SG.