Fixing track covariance matrix contents for muon/e-gamma smart slimming and PHYSLITE
This MR implements a change to the smart slimming lists for muons/e-gamma and the contents of PHYSLITE to reflect an EDM change where the diagonals and off-diagonals of the covariance matrix are stored separately. This was already done for InDetTrackParticles in !38992 but has not so far been changed for CombinedMuonTrackParticles, ExtrapolatedMuonTrackParticles, GSFTrackParticles, leading to reported crashes or only identity matrices being visible, depending on the analysis framework. This is corrected with this MR.