Drop unnecessary muon containers from trigger EDM
Dropping the 'OutsideIn' muon containers from the trigger EDM. These were originally added because of some problems saving the 'InsideOut' and 'OutsideIn' muons to the same container, but this has been fixed (quite a long time ago). Now the 'OutsideIn' muons are entirely a subset of the 'MuonsCB_RoI' container, and don't need to be duplicated. Also updating the monitoring to use the full final muon container as was originally intended.