MuonTrackSteeringTools: Remove garbage caching in MuPatCandidateTool.
Remove the slot-specific garbage-collection cache in MuPatCandidateTool.
Instead, declare garbage lists as local variables at the outermost interfaces,
and pass them as parameters.  In the process, update MuPatHit such that
it holds its parameter via a shareds_ptr (thus eliminating the need for
one of the garbage lists), and pass the broad measurement to it as a
unique_ptr.