Activation of new deadmap folder (ATLASRECTS-5014)
The deadmap format will be changed in rel.22.
The previous DB folder was /PIXEL/PixMapOverlay, and new folder /PIXEL/PixelModuleFeMask .
This new folder is not included in old global tag.
Thus, in the job option, the DB-tag is specified, but this should be removed when the q-test update new global tag.
Unfortunately, q221 uses old global tag, OFLCOND-MC16-SDR-25, which associates PixMapOverlay-SIM-MC16-000-02.
But the RUN-2 best-knowledge tag is based on PixMapOverlay-SIM-MC16-000-03, and new deadmap conditions also based on this tag (-03).
Therefore, the q221 test will not pass T0 policy. The reference file is updated.
On the other hand, the data q431 is no problem.