Made sure that the symbolic link creation could for sure happen.
In a weird race condition the python output directory may not exist
when the library was already built. Leading to a build error. And
while at it, I may as well have added a make_directory call for the
library output directory. Even though that should really always
exist by the time these commands are called...