Fix MuonTruthAssociationAlg for running without CSCs
The algorithm now checks to see if CSC hits are available before retrieving them, which was causing an exception which prevented the link from truth to reco muon being set.
This will solve the problem of muon efficiency (as determined by the MuonPerformanceAlg) appearing to be 0 for symmetrical Run-3 layouts, as reported in ATLASRECTS-5920.