MGC Protection against multiple generate calls
In some cases, it's necessary to call the generate function multiple
times from the same process (or at least it's allowed to do so, in
principle). With our reproducer-script setup, the data cards are backed
up before generation begins. We get a name conflict if this is done
multiple times, which can cause the job to crash. This avoids the
problem by including multiple backups of the cards (one per call to
generate), which should be even more reliable in reproducing jobs.