DataModelRoot: Try to work around apparent TClass race.
Sometimes when retrieving something with TClass, we seem to get back
something for which HasDictionary fails.  This is only in threaded
code and is not reproducible, so seems to be a race dealing with TClass
initialization.  As a workaround, try to retrieve the class a second
time in such cases.

See ATEAM-697.