CoreDumpSvc: add option to produce fast stack trace
Add the `CoreDumpSvc.FastStackTrace` option. If `True`, the
`CoreDumpSvc` will produce a stack trace of the current thread using the
stacktrace code from CxxUtils (formerly Seal). This can help in cases
when gdb does not manage to produce a stacktrace (e.g. due to timeouts
as in ATR-22583). Depending on the other configuration properties it
will then proceed with the "regular" (normally gdb) stacktrace
generation.