MuonMM_CnvTools: const fixes
Preparing to make methods of IMuonRdoToPrepDataTool const.

Adjust MmRdoToPrepDataTool* so that its interfaces may be made const.
Also enable thread-safety checking.
The old pre-MT MmRdoToPrepDataTool is declared as not thread-safe,
as it has some mutable state, while the mutable state is removed from
MmRdoToPrepDataToolMT.  MmRdoToPrepDataToolCore is now abstract
and is removed from the entries list.