tauRecTools: fix compatibility requirement between tau calo pt and substructure pt
Hello,

This MR is reverting !37684 . After thorough discussion, we agreed that the compatibility requirement, in terms of the number of sigmas between the calo pt and the substructure pt, is currently using an incorrect sigma definition.
The current "combined sigma" corresponds to the (parametrised) uncertainty on the weighted average according to the best linear unbiased estimate.
The original definition was instead estimating sigma(pt calo - pt substructure), which is used to quantify the compatibility of the 2 measurements.

For now I leave the "combined sigma" formula, although it is no longer used, we will come back to this code shortly to address the various FIX ME.

Cheers,
Bertrand