Adding new "Classifier" decoration for truth particles
PMG have recommended a new classification for truth particles. It's
already set in the truth particle classifier decoration algorithm, and
just needs to be handed off to the truth particle collections that are
made as a part of derivation making. This adds the Classification
decoration to the output of those collections.