Enable fast stack trace in CoreDumpSvc
Produce a "fast" stack trace by default in athena jobs. This initial
stacktrace will only report traces for the current thread but is much
faster than the full trace from gdb (which will follow). This can help in
cases when gdb does not manage to produce a stacktrace (e.g. due to
timeouts as in ATR-22583).