tauRecTools: add tau track flag to identify LRTs more easily
Hello,

This MR is adding a tau track flag to identify large radius tracks more easily, when LRTs are enabled in the tau reconstruction.
We can now retrieve specifically the LRTs associated with the tau via tau->tracks(xAOD::TauJetParameters::TauTrackFlag::LargeRadiusTrack).
This flag is agnostic w.r.t. the track classification, it is only used to tag LRTs, which will be useful e.g. to count LRTs within 0.4 of the tau axis.

As LRTs are not used by default, this MR does not change the reconstruction output.

Cheers,
Bertrand