jet hypo: check for missing hypo  scenarios
A hypo scenario is string whcih tells the jet condiguration what to do. It appears in the chain name.
The intention is that it is added to the a chainPart dict, a list of which is contained in the chan Dict.

There are to reasons why this could fail:

- the hypoScenario must be in the list SignatureDicts.JetChainPart[hypoScnario]. If it is not,
the hypoScneario will incorrectly take on the default value.

- the marker used in the chain name that triggers the creation of a new chainPart dict that with
contain the 'hypoScenario' key is "j0_". If this is missing, there will be no new chain part
and hence the hypo Scenario will not appear.

The current merge request prints out an error message if a hypoScenario is in the chain name,
but is not in a chainPart dictionary. For now, it takes no further action.