Add a nightly TrigInDetAnalysis test that changes the default trigger vertex algorithm
This adds the option to add a preinclude file for the job, rather than just a
post include.

This is necessary to change the vertex as the configuration needs to be steered differently
for the new vertex, it is not just a change in a parameter, but a change also in which
algorithms are used.