tauRecTools: fix FPE in TauSubstructureVariables
Hello,

This MR fixes divisions by 0 in the tau reconstruction, when ptDetectorAxis (not vertex-corrected) is null and ptIntermediateAxis (vertex-corrected) is not.
In the offline reconstruction we use the vertex correction everywhere, so it makes sense to use ptIntermediateAxis in the calculation of the ptRatioEflowApprox variable (the denominator approxSubstructure4Vec is also vertex corrected, for the cluster part).
This should address ATLASRECTS-5974 and ATLASRECTS-5925.
It will change the tau reconstruction to a very small extent, but this is desirable.

I took this opportunity to fix the code indentation and do some minor renaming.

Cheers,
Bertrand