jet hypo: remove raw print statements from production code.
uncontrolled print statements were producing unwanted output when run in
production. However, the agility and precision of such stastements are extemely
useful while developing code.

Printing is now done via DebugPrinter. This has a boolean flag to control whether
the printing is done. This is the same flage that produicess the chain diagnostic
information from the C++ code.