jet hypo - nmostly fixing code that handles identical Conditions
Previously, identacl conditions were recognised early on in the jet configuration. The approach
broke down when we introduced the constraint of one Condition per node.

The new approach is to build the single-Condition node tree, then for each node, check whether the
child nodes include identical nodes. If this is the case, all but one og the identical child nodes are deleted, and the multiplicity of the sole surviving child ndoe is set appropriately.

It turns out this also brings about code simplification.

This commit also includes improvments to code testing tools.