tauRecTools: fix FPE when clusters have null SECOND_LAMBDA moment
Hello,

This MR is adding a protection against pathological topoclusters that have a null SECOND_LAMBDA moment.
As discussed in ATLASRECTS-5974, this seems to happen when most of the cluster energy is in the presampler, and at |eta|~1.5.
These can't be good tau candidates, so an ad-hoc protection at the tau level is acceptable.

This MR does not change the reconstruction output except for taus with pathological clusters.

Hopefully, we now have very few sources of FPE left, if any...

The retrieval of decorations was improved a bit, and a new track variable was added to prepare the ground for the upcoming R22 tau ID and tau track classifier.

Cheers,
Bertrand