Fix for ATLASRECTS-5984
It's possible in rare cases for the combined fit to succeed with a track that has not been successfully extrapolated beforehand.
This is currently not correctly handled by the MuonCreatorTool: the muon is created, but the ME track is not added to it.
This simply fixes that issue, also removing the warning that is currently thrown.