tauRec: deploy tau thinning functionality
Hello,

This MR deploys the tau thinning functionality (ATLTAU-1770) in RecExCommon_topOptions (at the moment, it is in the cell thinning block).
The thinning is controlled by a new tauRec pt flag, which is currently set 0 not to activate the thinning yet.
It will be set to ~10 GeV for the R22 reprocessing.

Cheers,
Bertrand