jet hypo - provdie a new FastReduction jet grouper for a specific case.
Until now, a parent Condition was presented by all possible combinations of
jet groups passing child Conditions, with one jet group comming form each child,
provided that there was no repetition of an elemental jet (no sharing requirement).

Sometimes the full calculation of combinations can be avoided. In particular
if the parent is an AcceptAll condition, and it has one child. In this case
it is sufficient to calculate the union of the element jets used to form
the jet groups that pass the child Condition.

The njX (eg 10j40) are important examples of such a case.

The current code contains a JetGroupProduct factory to instantiate
either a JetGroupUnion to handle the special case, or a JetGroupProduct
otherwise. The factory is used from FastReducer::propagate_